requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
adafruit-circuitpython-pca9685==3.4.9
adafruit-circuitpython-servokit==1.3.15
RPi.GPIO==0.7.1
gpiozero==2.0
